requirements.txt
pandas==2.2.0
Pillow==10.2.0
Requests==2.31.0
stegano==0.11.3
